validações back-end e implentando criação de contatos

+++ api/src/app/controllers/ContactController.js
@@ -1,6 +1,7 @@
 const { response } = require("express");
 const req = require("express/lib/request");
 const ContactsRepository = require("../repositories/ContactsRepository");
+const isValidUUID = require('../utils/isValidUUID')
 
 class ContactController {
   async index(request, response) {
@@ -9,8 +10,14 @@ class ContactController {
 
     response.json(contacts);
   }
+
   async show(request, response) {
     const { id } = request.params;
+
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid contact id' })
+    }
+
     const contact = await ContactsRepository.findById(id);
     if (!contact) {
       //404: Not Found
@@ -24,18 +31,24 @@ class ContactController {
     if (!name) {
       return response.status(400).json({ error: "Name is required" });
     }
-    const contactExists = await ContactsRepository.findByEmail(email);
 
-    if (contactExists) {
-      return response
-        .status(400)
-        .json({ error: "This e-mail is already in use" });
+    if (category_id && !isValidUUID(category_id)) {
+      return response.status(400).json({ error: 'Invalid category' })
+    }
+
+    if (email) {
+      const contactExists = await ContactsRepository.findByEmail(email);
+      if (contactExists) {
+        return response
+          .status(400)
+          .json({ error: "This e-mail is already in use" });
+      }
     }
     const contact = await ContactsRepository.create({
       name,
-      email,
+      email: email || null,
       phone,
-      category_id,
+      category_id: category_id || null,
     });
     response.status(201).json(contact);
   }
@@ -45,26 +58,38 @@ class ContactController {
 
     const { name, email, phone, category_id } = request.body;
 
-    const contactExists = await ContactsRepository.findById(id);
-
-    if (!contactExists) {
-      response.status(404).json({ error: "Contact not found" });
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid contact id' })
     }
+
+    if (category_id && !isValidUUID(category_id)) {
+      return response.status(400).json({ error: 'Invalid category' })
+    }
+
+    const contactExists = await ContactsRepository.findById(id);
 
     if (!name) {
       response.status(400).json({ error: "Name is required" });
     }
 
-    const contactByEmail = await ContactsRepository.findByEmail(email);
-    if (contactByEmail && contactByEmail.id !== id) {
-      response.status(400).json({ error: "This e-mail is already in use." });
+    if (!contactExists) {
+      response.status(404).json({ error: "Contact not found" });
+    }
+
+
+
+    if (email) {
+      const contactByEmail = await ContactsRepository.findByEmail(email);
+      if (contactByEmail && contactByEmail.id !== id) {
+        response.status(400).json({ error: "This e-mail is already in use." });
+      }
     }
 
     const contact = await ContactsRepository.update(id, {
       name,
-      email,
+      email: email || null,
       phone,
-      category_id,
+      category_id: category_id || null,
     });
 
     response.json(contact);
@@ -72,6 +97,10 @@ class ContactController {
 
   async delete(request, response) {
     const { id } = request.params;
+
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid contact id' })
+    }
 
     await ContactsRepository.delete(id);
     //204: No Content
